+++ nbproject/project.properties
@@ -63,7 +63,7 @@ javadoc.splitindex=true
 javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
-main.class=ELLE_GUI.GUI
+main.class=ELLE_GUI.ellegui.GUI
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
