+++ src/ELLE_GUI/ellegui/ActionContainerBuilderFactory.java
@@ -1,5 +1,6 @@
-package ELLE_GUI;
+package ELLE_GUI.ellegui;
 
+import ELLE_GUI.ellegui.ActionBuilderHelper;
 import java.util.Collection;
 
 import javax.swing.Action;
