+++ src/ELLE_GUI/ellegui/ActionGroup.java
@@ -1,4 +1,4 @@
-package ELLE_GUI;
+package ELLE_GUI.ellegui;
 
 import java.awt.event.ActionEvent;
 import java.util.ArrayList;
