+++ src/ELLE_GUI/ellegui/CreateTables.java
@@ -1,4 +1,4 @@
-package ELLE_GUI;
+package ELLE_GUI.ellegui;
 
 import ELLE_GUI.presentation.filter.TableRowFilterSupport;
 import java.awt.Color;
