+++ src/ELLE_GUI/ellegui/DistinctColumnItem.java
@@ -1,4 +1,4 @@
-package ELLE_GUI;
+package ELLE_GUI.ellegui;
 
 import ELLE_GUI.presentation.filter.IValueWrapper;
 import java.math.BigDecimal;
