+++ src/ELLE_GUI/ellegui/EditDatabaseList.form
