+++ src/ELLE_GUI/ellegui/HeaderRenderer.java
@@ -1,4 +1,4 @@
-package ELLE_GUI;
+package ELLE_GUI.ellegui;
 
 import java.awt.Component;
