+++ src/ELLE_GUI/ellegui/JSearchTextField.java
@@ -1,4 +1,4 @@
-package ELLE_GUI;
+package ELLE_GUI.ellegui;
 
 import java.awt.Graphics;
 import java.awt.Image;
