+++ src/ELLE_GUI/ellegui/LogWindow.form
