+++ src/ELLE_GUI/ellegui/TradesView.java
@@ -1,4 +1,4 @@
-package ELLE_GUI;
+package ELLE_GUI.ellegui;
 
 public class TradesView {
